refactor(install): merge /installation-complete & /installation-slides
Ref: #1844
