+++ packages/ubuntu_desktop_installer/lib/pages.dart
@@ -2,8 +2,7 @@ export 'pages/active_directory/active_directory_page.dart';
 export 'pages/confirm/confirm_page.dart';
 export 'pages/filesystem/filesystem_page.dart';
 export 'pages/identity/identity_page.dart';
-export 'pages/installation_complete/installation_complete_page.dart';
-export 'pages/installation_slides/installation_slides_page.dart';
+export 'pages/install/install_page.dart';
 export 'pages/keyboard/keyboard_page.dart';
 export 'pages/locale/locale_page.dart';
 export 'pages/network/network_page.dart';
