+++ packages/ubuntu_desktop_installer/lib/pages/install/bottom_bar.dart
