+++ packages/ubuntu_desktop_installer/lib/pages/install/slide_transitions.dart
