--- packages/ubuntu_desktop_installer/lib/pages/installation_complete/installation_complete_model.dart
@@ -1,16 +1,0 @@
-import 'package:safe_change_notifier/safe_change_notifier.dart';
-import 'package:subiquity_client/subiquity_client.dart';
-import 'package:ubuntu_desktop_installer/services.dart';
-
-/// View model for [InstallationCompletePage].
-class InstallationCompleteModel extends SafeChangeNotifier {
-  /// Creates an instance with the given client.
-  InstallationCompleteModel(this._client, this._product);
-
-  final SubiquityClient _client;
-  final ProductService _product;
-
-  ProductInfo get productInfo => _product.getProductInfo();
-
-  Future<void> reboot() => _client.reboot(immediate: false);
-}
