+++ packages/ubuntu_desktop_installer/test/install/install_model_test.mocks.dart
@@ -1,9 +1,12 @@
 // Mocks generated by Mockito 5.3.2 from annotations
-// in ubuntu_desktop_installer/test/installation_complete/installation_complete_model_test.dart.
+// in ubuntu_desktop_installer/test/install/install_model_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
+import 'dart:async' as _i4;
+
 import 'package:mockito/mockito.dart' as _i1;
+import 'package:ubuntu_desktop_installer/services/journal_service.dart' as _i3;
 import 'package:ubuntu_desktop_installer/services/product_service.dart' as _i2;
 
 // ignore_for_file: type=lint
@@ -25,6 +28,29 @@ class _FakeProductInfo_0 extends _i1.SmartFake implements _i2.ProductInfo {
           parent,
           parentInvocation,
         );
+}
+
+/// A class which mocks [JournalService].
+///
+/// See the documentation for Mockito's code generation for more information.
+class MockJournalService extends _i1.Mock implements _i3.JournalService {
+  MockJournalService() {
+    _i1.throwOnMissingStub(this);
+  }
+
+  @override
+  _i4.Stream<String> start(
+    List<String>? ids, {
+    _i3.JournalOutput? output = _i3.JournalOutput.short,
+  }) =>
+      (super.noSuchMethod(
+        Invocation.method(
+          #start,
+          [ids],
+          {#output: output},
+        ),
+        returnValue: _i4.Stream<String>.empty(),
+      ) as _i4.Stream<String>);
 }
 
 /// A class which mocks [ProductService].
