+++ packages/ubuntu_desktop_installer/test/install/install_page_test.mocks.dart
@@ -1,5 +1,5 @@
 // Mocks generated by Mockito 5.3.2 from annotations
-// in ubuntu_desktop_installer/test/installation_slides/installation_slides_page_test.dart.
+// in ubuntu_desktop_installer/test/install/install_page_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
@@ -8,7 +8,7 @@ import 'dart:ui' as _i6;
 
 import 'package:flutter/widgets.dart' as _i5;
 import 'package:mockito/mockito.dart' as _i1;
-import 'package:ubuntu_desktop_installer/pages/installation_slides/installation_slides_model.dart'
+import 'package:ubuntu_desktop_installer/pages/install/install_model.dart'
     as _i3;
 import 'package:ubuntu_desktop_installer/services.dart' as _i2;
 
@@ -44,12 +44,11 @@ class _FakeInstallationEvent_1 extends _i1.SmartFake
         );
 }
 
-/// A class which mocks [InstallationSlidesModel].
+/// A class which mocks [InstallModel].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockInstallationSlidesModel extends _i1.Mock
-    implements _i3.InstallationSlidesModel {
-  MockInstallationSlidesModel() {
+class MockInstallModel extends _i1.Mock implements _i3.InstallModel {
+  MockInstallModel() {
     _i1.throwOnMissingStub(this);
   }
 
@@ -140,6 +139,15 @@ class MockInstallationSlidesModel extends _i1.Mock
         Invocation.method(
           #precacheSlideImages,
           [context],
+        ),
+        returnValue: _i4.Future<void>.value(),
+        returnValueForMissingStub: _i4.Future<void>.value(),
+      ) as _i4.Future<void>);
+  @override
+  _i4.Future<void> reboot() => (super.noSuchMethod(
+        Invocation.method(
+          #reboot,
+          [],
         ),
         returnValue: _i4.Future<void>.value(),
         returnValueForMissingStub: _i4.Future<void>.value(),
