+++ packages/ubuntu_desktop_installer/test/install/slides_test.dart
