+++ packages/ubuntu_desktop_installer/test/install/slides_test.mocks.dart
@@ -1,5 +1,5 @@
 // Mocks generated by Mockito 5.3.2 from annotations
-// in ubuntu_desktop_installer/test/installation_slides/slides_test.dart.
+// in ubuntu_desktop_installer/test/install/slides_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
